refactor: 코드 리팩토링 및 불필요 import 제거
