+++ src/main/java/com/threestar/trainus/domain/metadata/repository/ProfileMetadataRepository.java
@@ -1,4 +1,4 @@
-package com.threestar.trainus.domain.metadata.repositroy;
+package com.threestar.trainus.domain.metadata.repository;
 
 import java.util.Optional;
