+++ src/main/java/com/threestar/trainus/domain/profile/mapper/ProfileMapper.java
@@ -1,9 +1,6 @@
 package com.threestar.trainus.domain.profile.mapper;
 
-import com.threestar.trainus.domain.metadata.dto.ProfileMetadataResponseDto;
-import com.threestar.trainus.domain.profile.dto.ProfileDetailResponseDto;
 import com.threestar.trainus.domain.profile.dto.ProfileResponseDto;
-import com.threestar.trainus.domain.profile.dto.ProfileUpdateRequestDto;
 import com.threestar.trainus.domain.profile.entity.Profile;
 import com.threestar.trainus.domain.user.entity.User;
 
@@ -20,15 +17,6 @@ public class ProfileMapper {
 			profile.getProfileImage(),
 			profile.getIntro()
 		);
-	}
-
-	public static Profile toUpdateEntity(Profile profile, ProfileUpdateRequestDto requestDto) {
-		return Profile.builder()
-			.id(profile.getId())
-			.user(profile.getUser())
-			.profileImage(requestDto.profileImage())
-			.intro(requestDto.intro())
-			.build();
 	}
 
 	public static Profile toDefaultEntity(User user) {
